+++ optimum/intel/openvino/modeling_base.py
@@ -52,6 +52,7 @@ class OVBaseModel(OptimizedModel):
     auto_model_class = None
     export_feature = None
     _supports_cache_class = False
+    _library_name = "transformers"
 
     def __init__(
         self,
@@ -452,6 +453,7 @@ class OVBaseModel(OptimizedModel):
             force_download=force_download,
             trust_remote_code=trust_remote_code,
             ov_config=ov_config,
+            library_name=cls.library_name,
         )
 
         config.save_pretrained(save_dir_path)
